requirements.txt
pandas
requests
python-decouple
openpyxl
streamlit
